+++ xbmc/cores/VideoPlayer/VideoPlayer.h
@@ -521,7 +521,7 @@ protected:
   {
     double  lastpts;  // holds last display pts during ff/rw operations
     int64_t lasttime;
-    int lastseekpts;
+    double lastseekpts;
     double  lastabstime;
   } m_SpeedState;
   std::atomic_bool m_canTempo;
